Move SplitData and SplitterMessageFilter to separate files
